Fixed some missing types and some types moved to types folder.
