+++ src/pages/Registration/index.tsx
@@ -31,24 +31,10 @@ import FileOperations from '../../utils/FileOperations';
 
 // Type imports
 import { IDialogState } from '../../types/globals/Dialog';
-import { picturePath } from '../../types/pages/Registration';
+import { picturePath, ISignUpForm } from '../../types/pages/Registration';
 
 // Theme imports.
 import { styles, styledComponents } from './styles';
-
-// Interface declarations.
-interface ISignUpForm {
-  fullName: string,
-  birthDate: Date | null,
-  email: string,
-  state: string,
-  city: string,
-  address: string,
-  phoneNumber: string,
-  username: string,
-  password: string,
-  passwordConfirmation: string,
-}
 
 // Component.
 export default function Registration() : JSX.Element {
