+++ src/types/components/AnimalCard/index.ts
@@ -1,4 +1,5 @@
 import { ReactNode } from 'react';
+import { ViewProps } from 'react-native';
 
 export interface IAnimalCardProps {
   title: string;
@@ -7,4 +8,8 @@ export interface IAnimalCardProps {
   body?: ReactNode;
   headerBackground: string;
   onPress?: () => void;
+}
+
+export interface CardHeaderProps extends ViewProps {
+  backgroundColor: string;
 }
