+++ src/types/pages/Registration/index.ts
@@ -1,1 +1,14 @@
 export type picturePath = string | null;
+
+export interface ISignUpForm {
+    fullName: string,
+    birthDate: Date | null,
+    email: string,
+    state: string,
+    city: string,
+    address: string,
+    phoneNumber: string,
+    username: string,
+    password: string,
+    passwordConfirmation: string,
+}
